putnbr updated to be usefull only for ints

+++ ft_printf/ft_putnbr.c
@@ -6,14 +6,14 @@
 /*   By: rita <rita@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/29 13:34:55 by fwebe-ir          #+#    #+#             */
-/*   Updated: 2024/11/30 12:31:29 by rita             ###   ########.fr       */
+/*   Updated: 2024/12/02 10:02:51 by rita             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
 //TODO should return double instead of int like that you handle ints and doubles.
-void	ft_putnbr(double n)
+void	ft_putnbr(int n)
 {
 	char	c;
 
@@ -30,6 +30,6 @@ void	ft_putnbr(double n)
 	{
 		ft_putnbr(n / 10);
 	}
-	c = ((int)n % 10) + '0';
+	c = (n % 10) + '0';
 	write(1, &c, 1);
 }
